feat(admin): fit countdown preview, MLH badge preview, modular tabs + site/ folder
- Misc tab live preview: countdown now scales down to fit the panel
  (new ScaledPreview measures natural size and transforms) instead of
  horizontal scrolling; MLH badge previewed with the real badge art —
  grayscale/dimmed when off, full color when on
- badge URL/alt shared between Navbar and admin via lib/mlh.js
- split the four large admin tabs into per-tab modules:
  team/ sponsors/ schedule/ gallery/ (tab + modals + panels + utils)
- shared admin building blocks: icons.jsx, useObjectUrls hook, and ui.jsx
  additions (Toggle, CollapseTitle, CardOverlay, CardMoveButtons,
  ScaledPreview)
- public site components reorganized into components/site/
