+++ frontend/src/components/site/FAQ.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { motion, AnimatePresence } from 'motion/react';
-import { faqs } from '../data/faq';
+import { faqs } from '../../data/faq';
 
 function FAQItem({ faq }) { // this is the FAQ item component
   const [isOpen, setIsOpen] = useState(false); // this state is used to store whether the FAQ item is open or not
