+++ frontend/src/components/site/Slideshow.jsx
