+++ frontend/src/components/site/SponsorsCarousel.jsx
@@ -5,7 +5,7 @@
 
 import { Link } from 'react-router-dom';            // Link for navigating to /sponsors
 import { motion } from 'motion/react';              // Framer Motion for infinite carousel loop
-import { useSponsors } from '../hooks/useSponsors';  // fetches sponsors from the API (static fallback)
+import { useSponsors } from '../../hooks/useSponsors';  // fetches sponsors from the API (static fallback)
 
 export default function SponsorsCarousel() {
   const { sponsors } = useSponsors();
