+++ frontend/src/lib/mlh.js
@@ -1,0 +1,10 @@
+// Official MLH trust badge assets — used by the public navbar and the
+// admin Misc tab preview so the two can never drift apart.
+
+export const MLH_BADGE_HREF =
+  "https://mlh.io/na?utm_source=na-hackathon&utm_medium=TrustBadge&utm_campaign=2026-season&utm_content=gray";
+
+export const MLH_BADGE_SRC =
+  "https://s3.amazonaws.com/logged-assets/trust-badge/2026/mlh-trust-badge-2026-gray.svg";
+
+export const MLH_BADGE_ALT = "Major League Hacking 2026 Hackathon Season";
