+++ frontend/src/styles/components.css
@@ -32,10 +32,9 @@
 
   .card-sponsor {
     @apply transition-all duration-300 ease-brand hover:-translate-y-1
-           flex items-center justify-center p-4 sm:p-6
-           grayscale hover:grayscale-0;
+           flex items-center justify-center p-4 sm:p-6;
     &:hover {
-      filter: grayscale(0) drop-shadow(0 0 14px rgba(255, 255, 255, 0.55));
+      filter: drop-shadow(0 0 14px rgba(255, 255, 255, 0.55));
     }
   }
 
@@ -78,7 +77,7 @@
     @apply bg-surface border border-border rounded-card shadow-card
            flex flex-col items-center justify-center gap-4 p-6
            transition-all duration-300 ease-brand
-           hover:-translate-y-1 grayscale hover:grayscale-0 hover:shadow-glow hover:border-ultraviolet/40;
+           hover:-translate-y-1 hover:shadow-glow hover:border-ultraviolet/40;
   }
 
   /* Tier colour accents on hover */
